Implements CompilerHandler and unit tests. (#13)
* Adds Status Enum for jobs, feat #4
Adds Status Enum that can be used for compiling and testing. The enum contains
PENDING, which is used when the process is started,
SUCCESS, which is used when the process is successful,
FAILURE, which is used when the process finished but is unsuccessful,
ERROR, which is used when the proccess throws an exception.

* removed from cache since they are in the .gitignore

* Implements CompileHandler and unit tests, fix #4

Implements CompileHandler which is responsible for compiling a commit in
a cloned repo. The unit tests tests if the CompileHandler successfully can compile a given project
and if the privte method convertInputStreamToString works as expected.

* Adds unit test that checks compilation failure

Adds a unit test that pulls a branch (test/build-fail) which contains code with syntax errors,
thus the code should not be able to build. The test is passed if the build is failed.